+++ tests/end2end/test_adapter.py
@@ -32,8 +32,8 @@ from zhmcclient.testutils import hmc_definition, hmc_session  # noqa: F401, E501
 from zhmcclient.testutils import dpm_mode_cpcs  # noqa: F401, E501
 # pylint: enable=line-too-long,unused-import
 
-from .utils import pick_test_resources, runtest_find_list, TEST_PREFIX, \
-    skip_warn, standard_partition_props
+from .utils import skip_warn, pick_test_resources, TEST_PREFIX, \
+    standard_partition_props, runtest_find_list, runtest_get_properties
 
 urllib3.disable_warnings()
 
@@ -76,6 +76,39 @@ def test_adapter_find_list(dpm_mode_cpcs):  # noqa: F811
                 session, cpc.adapters, adapter.name, 'name', 'object-uri',
                 ADAPTER_VOLATILE_PROPS, ADAPTER_MINIMAL_PROPS,
                 ADAPTER_LIST_PROPS)
+
+
+def test_adapter_property(dpm_mode_cpcs):  # noqa: F811
+    # pylint: disable=redefined-outer-name
+    """
+    Test property related methods
+    """
+    if not dpm_mode_cpcs:
+        pytest.skip("HMC definition does not include any CPCs in DPM mode")
+
+    for cpc in dpm_mode_cpcs:
+        assert cpc.dpm_enabled
+
+        client = cpc.manager.client
+        session = cpc.manager.session
+        hd = session.hmc_definition
+
+        # Pick the adapters to test with
+        adapter_list = cpc.adapters.list()
+        if not adapter_list:
+            skip_warn("No adapters on CPC {c} managed by HMC {h}".
+                      format(c=cpc.name, h=hd.host))
+        adapter_list = pick_test_resources(adapter_list)
+
+        for adapter in adapter_list:
+            print("Testing on CPC {c} with adapter {a!r}".
+                  format(c=cpc.name, a=adapter.name))
+
+            # Select a property that is not returned by list()
+            non_list_prop = 'description'
+
+            runtest_get_properties(
+                client, adapter.manager, non_list_prop, None)
 
 
 def test_adapter_hs_crud(dpm_mode_cpcs):  # noqa: F811