+++ tests/end2end/test_ldap_server_definition.py
@@ -30,8 +30,8 @@ import zhmcclient
 from zhmcclient.testutils import hmc_definition, hmc_session  # noqa: F401, E501
 # pylint: enable=line-too-long,unused-import
 
-from .utils import pick_test_resources, runtest_find_list, TEST_PREFIX, \
-    skip_warn
+from .utils import skip_warn, pick_test_resources, TEST_PREFIX, \
+    runtest_find_list, runtest_get_properties
 
 urllib3.disable_warnings()
 
@@ -77,6 +77,39 @@ def test_ldapsrvdef_find_list(hmc_session):  # noqa: F811
             hmc_session, console.ldap_server_definitions, ldapsrvdef.name,
             'name', 'element-uri', LDAPSRVDEF_VOLATILE_PROPS,
             LDAPSRVDEF_MINIMAL_PROPS, LDAPSRVDEF_LIST_PROPS)
+
+
+def test_ldapsrvdef_property(hmc_session):  # noqa: F811
+    # pylint: disable=redefined-outer-name
+    """
+    Test property related methods
+    """
+    client = zhmcclient.Client(hmc_session)
+    console = client.consoles.console
+    hd = hmc_session.hmc_definition
+
+    api_version = client.query_api_version()
+    hmc_version = api_version['hmc-version']
+    hmc_version_info = tuple(map(int, hmc_version.split('.')))
+    if hmc_version_info < (2, 13, 0):
+        skip_warn("HMC {h} of version {v} does not yet support LDAP server "
+                  "definitions".format(h=hd.host, v=hmc_version))
+
+    # Pick the LDAP server definitions to test with
+    ldapsrvdef_list = console.ldap_server_definitions.list()
+    if not ldapsrvdef_list:
+        skip_warn("No LDAP server definitions defined on HMC {h}".
+                  format(h=hd.host))
+    ldapsrvdef_list = pick_test_resources(ldapsrvdef_list)
+
+    for ldapsrvdef in ldapsrvdef_list:
+        print("Testing with LDAP server definition {d!r}".
+              format(d=ldapsrvdef.name))
+
+        # Select a property that is not returned by list()
+        non_list_prop = 'description'
+
+        runtest_get_properties(client, ldapsrvdef.manager, non_list_prop, None)
 
 
 def test_ldapsrvdef_crud(hmc_session):  # noqa: F811