+++ zhmcclient/_cpc.py
@@ -113,7 +113,8 @@ class CpcManager(BaseManager):
             oid_prop='object-id',
             uri_prop='object-uri',
             name_prop='name',
-            query_props=query_props)
+            query_props=query_props,
+            supports_properties=True)
         self._client = client
 
     @property